# Elide assertValid & assertValidParts in production