@decentralion 2018-12-02
I'm not sure if it disappears entirely in production. It's hard to tell with my instrumentation because all the function names are lost.