@wchargin 2018-12-02
From the minified-then-prettified code (comments are my own):

```js
        function t(e, t) {  // assertValid
          return;
        }
        function r(e, t) {  // assertValidParts
          return;
        }
        function o(e) {  // fromParts
          return r(e), h + y + i(e);
        }
        function u(e) {  // toParts
          t(e);
          var r = e.split(y);
          return r.slice(1, r.length - 1);
        }
        function f(e, r) {  // hasPrefix
          return t(e, "address"), t(r, "prefix"), e.startsWith(r);
        }
```

So it looks like the functions continue to exist (which they must,
because they’re part of the `AddressModule` interface and thus part of
the return value) with now-empty bodies, and the call sites also
continue to exist.

An optimizer _should_ be able to remove the call sites at least in
`fromParts` and friends (maybe not in more distant call sites in other
modules, where the indirection is through a module object: it should
still be valid because we `Object.freeze` the module, but this is far
less obvious). But it also probably doesn’t matter very much, because
any decent JIT will likely optimize these call sites away, too.